requirements.txt
bip-utils==2.3.0
django==4.0.4
anchorpy==0.15.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
dj-database-url==0.5.0
django-celery-results==2.3.1
django-bmemcached==0.3.0
django-cors-headers==3.12.0
django-encrypted-model-fields==0.6.5
django-filter==21.1
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
drf-yasg==1.20.0
ed25519==1.5
eth-account==0.9.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
locust==2.15.1
redis==4.3.3
python-dotenv==0.20.0
python-memcached==1.59
sentry-sdk==1.5.12
uWSGI==2.0.22
web3==6.9.0
whitenoise==6.1.0
solana==0.29.1
solders==0.14.3
coverage~=7.3.2
pytz~=2023.3.post1
requests~=2.31.0
celery~=5.3.4
